requirements.txt
Flask==3.0.3
boto3==1.34.0
pandas==2.2.2
pandasql==0.7.3
python-dotenv==1.0.1
